finance_app/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Django==5.1.7
django-environ==0.12.0
django-ratelimit==4.1.0
django-sslserver==0.22
django-sslserver-v2==1.0
django-widget-tweaks==1.5.0
dotmap==1.3.30
frozendict==2.4.6
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multitasking==0.0.11
numpy==2.2.4
orjson==3.10.15
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
quantcrypt==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typer==0.15.2
types-zxcvbn==4.5.0.20250223
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
yfinance==0.2.55
zxcvbn==4.5.0
